requirements.txt
huey
redispy
